// frontend/generated/models/ConnectionInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ConnectionInfo = {
    server_real_ip?: string;
    user_ip?: string;
    token?: string;
    fromdomain?: string;
    worktime?: string;
    worktime_tmp?: string;
    request_uuid?: string;
};